Ignore build artifact generated from README directions
Running the installation directions in the README generated egg-info
artifacts, which made the repository appear to be in a "Dirty" state due
to untracked, not-ignored content.  This is a mild point of confusion
for other repositories that would want to track this repository as a
submodule.  I suggest ignoring this build artifact that would appear
from regular usage as prescribed in the README.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>